+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * `selected_function` and `function_lists_file` in gseapy_orsum.
 * Use of function table as input for `esmecata_gseapy gseapy_enrichr` (for example to be used on PICRUSt2 results).
 * Use of KEGG Orthologs in `esmecata_gseapy gseapy_enrichr` with `--ko` parameter.
+* `--annot-names` parameter to give a json dict for annotation names. Default downloads annotation names from EC/GO/KO.
 * Tests for function `create_dataset_annotation_file`.
 
 ## Fix
@@ -15,6 +16,7 @@
 
 ## Modify
 
+* Remove old parameters to download EC and GO names in `esmecata_gseapy gseapy_enrichr`.
 * Update variable names in `gseapy_orsum.py`.
 * Update readme on enrichment analysis.